requirements/base.txt
django==2.0.1
django-money==0.12.3
coverage==4.4.2
psycopg2==2.7.3.2
django-bootstrap-form==3.3
reportlab==3.4.0
https://github.com/johngriebel/django-address/archive/master.zip
dj_database_url==0.4.2
django-rest-framework==3.7.7
drf-nested-routers==0.90.2
